+++ primebot/config/config_loader.py
@@ -2,7 +2,7 @@ import yaml
 from pymongo import MongoClient
 from primebot.utils import gazelle
 
-__all__ = ['conf', 'db', 'sows']
+__all__ = ['conf', 'db', 'sows', 'ops']
 
 
 def load_config():
@@ -13,4 +13,8 @@ def load_config():
 
 conf = load_config()
 db = MongoClient(conf['mongo_db']).primebot
-sows = gazelle.WhatAPI(conf['sows']['username'], conf['sows']['password'])
+sows = sows.WhatAPI(conf['sows']['username'], conf['sows']['password'])
+
+conf = load_config()
+db = MongoClient(conf['mongo_db']).primebot
+ops = ops.WhatAPI(conf['ops']['username'], conf['ops']['password'])
